APIs/googleapis.com/datacatalog/v1beta1/python-aiohttp-client/tests/test_projects_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.get_iam_policy_request import GetIamPolicyRequest
from openapi_server.models.google_cloud_datacatalog_v1beta1_entry import GoogleCloudDatacatalogV1beta1Entry
from openapi_server.models.google_cloud_datacatalog_v1beta1_entry_group import GoogleCloudDatacatalogV1beta1EntryGroup
from openapi_server.models.google_cloud_datacatalog_v1beta1_export_taxonomies_response import GoogleCloudDatacatalogV1beta1ExportTaxonomiesResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_import_taxonomies_request import GoogleCloudDatacatalogV1beta1ImportTaxonomiesRequest
from openapi_server.models.google_cloud_datacatalog_v1beta1_import_taxonomies_response import GoogleCloudDatacatalogV1beta1ImportTaxonomiesResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_list_entries_response import GoogleCloudDatacatalogV1beta1ListEntriesResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_list_entry_groups_response import GoogleCloudDatacatalogV1beta1ListEntryGroupsResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_list_policy_tags_response import GoogleCloudDatacatalogV1beta1ListPolicyTagsResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_list_tags_response import GoogleCloudDatacatalogV1beta1ListTagsResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_list_taxonomies_response import GoogleCloudDatacatalogV1beta1ListTaxonomiesResponse
from openapi_server.models.google_cloud_datacatalog_v1beta1_policy_tag import GoogleCloudDatacatalogV1beta1PolicyTag
from openapi_server.models.google_cloud_datacatalog_v1beta1_rename_tag_template_field_enum_value_request import GoogleCloudDatacatalogV1beta1RenameTagTemplateFieldEnumValueRequest
from openapi_server.models.google_cloud_datacatalog_v1beta1_tag import GoogleCloudDatacatalogV1beta1Tag
from openapi_server.models.google_cloud_datacatalog_v1beta1_tag_template import GoogleCloudDatacatalogV1beta1TagTemplate
from openapi_server.models.google_cloud_datacatalog_v1beta1_tag_template_field import GoogleCloudDatacatalogV1beta1TagTemplateField
from openapi_server.models.google_cloud_datacatalog_v1beta1_taxonomy import GoogleCloudDatacatalogV1beta1Taxonomy
from openapi_server.models.policy import Policy
from openapi_server.models.set_iam_policy_request import SetIamPolicyRequest
from openapi_server.models.test_iam_permissions_request import TestIamPermissionsRequest
from openapi_server.models.test_iam_permissions_response import TestIamPermissionsResponse


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_entry_groups_create(client):
    """Test case for datacatalog_projects_locations_entry_groups_create

    
    """
    body = {"dataCatalogTimestamps":{"expireTime":"expireTime","createTime":"createTime","updateTime":"updateTime"},"displayName":"displayName","name":"name","description":"description"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('entryGroupId', 'entry_group_id_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/entryGroups'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_entry_groups_entries_create(client):
    """Test case for datacatalog_projects_locations_entry_groups_entries_create

    
    """
    body = {"schema":{"columns":[{"mode":"mode","column":"column","description":"description","type":"type","subcolumns":[null,null]},{"mode":"mode","column":"column","description":"description","type":"type","subcolumns":[null,null]}]},"displayName":"displayName","gcsFilesetSpec":{"filePatterns":["filePatterns","filePatterns"],"sampleGcsFileSpecs":[{"gcsTimestamps":{"expireTime":"expireTime","createTime":"createTime","updateTime":"updateTime"},"filePath":"filePath","sizeBytes":"sizeBytes"},{"gcsTimestamps":{"expireTime":"expireTime","createTime":"createTime","updateTime":"updateTime"},"filePath":"filePath","sizeBytes":"sizeBytes"}]},"sourceSystemTimestamps":{"expireTime":"expireTime","createTime":"createTime","updateTime":"updateTime"},"description":"description","type":"ENTRY_TYPE_UNSPECIFIED","userSpecifiedSystem":"userSpecifiedSystem","linkedResource":"linkedResource","usageSignal":{"usageWithinTimeRange":{"key":{"totalFailures":5.962134,"totalCancellations":0.8008282,"totalExecutionTimeForCompletionsMillis":1.4658129,"totalCompletions":6.0274563}},"updateTime":"updateTime"},"userSpecifiedType":"userSpecifiedType","name":"name","integratedSystem":"INTEGRATED_SYSTEM_UNSPECIFIED","bigqueryTableSpec":{"tableSpec":{"groupedEntry":"groupedEntry"},"viewSpec":{"viewQuery":"viewQuery"},"tableSourceType":"TABLE_SOURCE_TYPE_UNSPECIFIED"},"bigqueryDateShardedSpec":{"tablePrefix":"tablePrefix","dataset":"dataset","shardCount":"shardCount"}}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('entryId', 'entry_id_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/entries'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_entry_groups_entries_list(client):
    """Test case for datacatalog_projects_locations_entry_groups_entries_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('pageSize', 56),
                    ('pageToken', 'page_token_example'),
                    ('readMask', 'read_mask_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{parent}/entries'.format(parent='parent_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_entry_groups_list(client):
    """Test case for datacatalog_projects_locations_entry_groups_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('pageSize', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{parent}/entryGroups'.format(parent='parent_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_entry_groups_tags_create(client):
    """Test case for datacatalog_projects_locations_entry_groups_tags_create

    
    """
    body = {"template":"template","column":"column","name":"name","fields":{"key":{"stringValue":"stringValue","enumValue":{"displayName":"displayName"},"displayName":"displayName","timestampValue":"timestampValue","boolValue":True,"doubleValue":0.8008281904610115,"order":6}},"templateDisplayName":"templateDisplayName"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/tags'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_entry_groups_tags_list(client):
    """Test case for datacatalog_projects_locations_entry_groups_tags_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('pageSize', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{parent}/tags'.format(parent='parent_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_tag_templates_create(client):
    """Test case for datacatalog_projects_locations_tag_templates_create

    
    """
    body = {"displayName":"displayName","name":"name","fields":{"key":{"isRequired":True,"displayName":"displayName","name":"name","description":"description","type":{"enumType":{"allowedValues":[{"displayName":"displayName"},{"displayName":"displayName"}]},"primitiveType":"PRIMITIVE_TYPE_UNSPECIFIED"},"order":0}}}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('tagTemplateId', 'tag_template_id_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/tagTemplates'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_tag_templates_fields_create(client):
    """Test case for datacatalog_projects_locations_tag_templates_fields_create

    
    """
    body = {"isRequired":True,"displayName":"displayName","name":"name","description":"description","type":{"enumType":{"allowedValues":[{"displayName":"displayName"},{"displayName":"displayName"}]},"primitiveType":"PRIMITIVE_TYPE_UNSPECIFIED"},"order":0}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('tagTemplateFieldId', 'tag_template_field_id_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/fields'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_tag_templates_fields_enum_values_rename(client):
    """Test case for datacatalog_projects_locations_tag_templates_fields_enum_values_rename

    
    """
    body = {"newEnumValueDisplayName":"newEnumValueDisplayName"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{namerenam}'.format(name='name_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_create(client):
    """Test case for datacatalog_projects_locations_taxonomies_create

    
    """
    body = {"activatedPolicyTypes":["POLICY_TYPE_UNSPECIFIED","POLICY_TYPE_UNSPECIFIED"],"displayName":"displayName","service":{"identity":"identity","name":"MANAGING_SYSTEM_UNSPECIFIED"},"name":"name","description":"description","policyTagCount":0,"taxonomyTimestamps":{"expireTime":"expireTime","createTime":"createTime","updateTime":"updateTime"}}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/taxonomies'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_export(client):
    """Test case for datacatalog_projects_locations_taxonomies_export

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('serializedTaxonomies', True),
                    ('taxonomies', ['taxonomies_example'])]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{parent}/taxonomies:export'.format(parent='parent_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_import(client):
    """Test case for datacatalog_projects_locations_taxonomies_import

    
    """
    body = {"inlineSource":{"taxonomies":[{"policyTags":[{"childPolicyTags":[null,null],"policyTag":"policyTag","displayName":"displayName","description":"description"},{"childPolicyTags":[null,null],"policyTag":"policyTag","displayName":"displayName","description":"description"}],"activatedPolicyTypes":["POLICY_TYPE_UNSPECIFIED","POLICY_TYPE_UNSPECIFIED"],"displayName":"displayName","description":"description"},{"policyTags":[{"childPolicyTags":[null,null],"policyTag":"policyTag","displayName":"displayName","description":"description"},{"childPolicyTags":[null,null],"policyTag":"policyTag","displayName":"displayName","description":"description"}],"activatedPolicyTypes":["POLICY_TYPE_UNSPECIFIED","POLICY_TYPE_UNSPECIFIED"],"displayName":"displayName","description":"description"}]}}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/taxonomies:import'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_list(client):
    """Test case for datacatalog_projects_locations_taxonomies_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('filter', 'filter_example'),
                    ('pageSize', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{parent}/taxonomies'.format(parent='parent_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_create(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_create

    
    """
    body = {"childPolicyTags":["childPolicyTags","childPolicyTags"],"displayName":"displayName","name":"name","description":"description","parentPolicyTag":"parentPolicyTag"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{parent}/policyTags'.format(parent='parent_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_delete(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_delete

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('force', True)]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/v1beta1/{name}'.format(name='name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_get(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_get

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('readMask', 'read_mask_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{name}'.format(name='name_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_get_iam_policy(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_get_iam_policy

    
    """
    body = {"options":{"requestedPolicyVersion":0}}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{resourceget_iam_polic}'.format(resource='resource_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_list(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_list

    
    """
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('pageSize', 56),
                    ('pageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1beta1/{parent}/policyTags'.format(parent='parent_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_patch(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_patch

    
    """
    body = {"childPolicyTags":["childPolicyTags","childPolicyTags"],"displayName":"displayName","name":"name","description":"description","parentPolicyTag":"parentPolicyTag"}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example'),
                    ('updateMask', 'update_mask_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PATCH',
        path='/v1beta1/{name}'.format(name='name_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_set_iam_policy(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_set_iam_policy

    
    """
    body = {"policy":{"bindings":[{"condition":{"expression":"expression","description":"description","location":"location","title":"title"},"role":"role","members":["members","members"]},{"condition":{"expression":"expression","description":"description","location":"location","title":"title"},"role":"role","members":["members","members"]}],"etag":"etag","version":0}}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{resourceset_iam_polic}'.format(resource='resource_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_datacatalog_projects_locations_taxonomies_policy_tags_test_iam_permissions(client):
    """Test case for datacatalog_projects_locations_taxonomies_policy_tags_test_iam_permissions

    
    """
    body = {"permissions":["permissions","permissions"]}
    params = [('$.xgafv', 'xgafv_example'),
                    ('access_token', 'access_token_example'),
                    ('alt', 'alt_example'),
                    ('callback', 'param_callback_example'),
                    ('fields', 'fields_example'),
                    ('key', 'key_example'),
                    ('oauth_token', 'oauth_token_example'),
                    ('prettyPrint', True),
                    ('quotaUser', 'quota_user_example'),
                    ('upload_protocol', 'upload_protocol_example'),
                    ('uploadType', 'upload_type_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='POST',
        path='/v1beta1/{resourcetest_iam_permission}'.format(resource='resource_example'),
        headers=headers,
        json=body,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

